resize_images/requirements.txt
flask==2.0.1
Pillow==9.0.0 